01-function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-keyvault
azure-identity
cx-Oracle==8.3.0
